parser file renamed to serial-port-parser
