+++ packages/pure/src/device/serial-port-parser/serial-port-parser.ts
